STL work
Did some random cleanup of old Hy names.
Wanted to use a very basic amount of STL (pretty much only stuff that happens at compile time, not any of the memory allocation structures), and had to fix a whole bunch of clashing errors between the Windows Kit STL and the MSVC STL.
All this because I got tired of seeing the little yellow line complaining that I was logging structures that I knew were just bitfields the exact size of an INT (and one odd one that is actually the size of a UINT64, but that's beside the point, this code still handles that correctly).